Update app structure
- Restructuring app
