--- timer/Coordinator/ProductivityViewCoordinator.swift
@@ -1,44 +1,0 @@
-//
-//  ProductivityViewCoordinator.swift
-//  timer
-//
-//  Created by Jeong Jin Eun on 09/04/2019.
-//  Copyright © 2019 Jeong Jin Eun. All rights reserved.
-//
-
-/// Route from one touch timer view
-class ProductivityViewCoordinator: CoordinatorProtocol {
-     // MARK: route enumeration
-    enum ProductivityRoute {
-        case createTimerSet(TimeSet)
-    }
-
-    // MARK: properties
-    weak var rootViewController: ProductivityViewController!
-    let provider: ServiceProviderProtocol
-    
-    required init(provider: ServiceProviderProtocol, rootViewController: ProductivityViewController) {
-        self.provider = provider
-        self.rootViewController = rootViewController
-    }
-    
-    func present(for route: ProductivityRoute) {
-        switch route {
-        case let .createTimerSet(timeSet):
-            Logger.verbose("presenting create timer set view controller.")
-            
-            let viewController = CreateTimerSetViewController()
-            let coordinator = CreateTimerSetViewCoordinator(provider: provider, rootViewController: viewController)
-            
-            // DI
-            viewController.coordinator = coordinator
-            viewController.reactor = CreateTimerSetViewReactor(timerService: provider.timerService, timeSet: timeSet)
-            
-            // Hide tab bar when enter view controller
-            viewController.hidesBottomBarWhenPushed = true
-            
-            // Push view controller
-            rootViewController.navigationController?.pushViewController(viewController, animated: true)
-        }
-    }
-}
