+++ timer/Model/Base/BaseTableItem.swift
