+++ timer/Module/AppInfo/AppInfoViewReactor.swift
