+++ timer/Module/Main/LocalTimeSet/LocalTimeSetView.swift
@@ -1,0 +1,22 @@
+//
+//  LocalTimeSetView.swfit
+//  timer
+//
+//  Created by Jeong Jin Eun on 09/04/2019.
+//  Copyright © 2019 Jeong Jin Eun. All rights reserved.
+//
+
+import UIKit
+
+class LocalTimeSetView: UIView {
+    // MARK: - view properties
+    
+    // MARK: - constructor
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
