+++ timer/Module/Main/LocalTimeSet/LocalTimeSetViewController.swift
@@ -1,26 +1,26 @@
 //
-//  CreateTimerSetViewController.swift
+//  LocalTimeSetViewController.swift
 //  timer
 //
-//  Created by JSilver on 19/05/2019.
+//  Created by Jeong Jin Eun on 09/04/2019.
 //  Copyright © 2019 Jeong Jin Eun. All rights reserved.
 //
 
-import UIKit
 import RxSwift
 import RxCocoa
 import ReactorKit
+import RxDataSources
 
-class CreateTimerSetViewController: BaseViewController, View {
+class LocalTimeSetViewController: BaseViewController, View {
     // MARK: - view properties
-    private var createTimerSetView: CreateTimerSetView { return self.view as! CreateTimerSetView }
+    private var timeSetView: LocalTimeSetView { return self.view as! LocalTimeSetView }
     
     // MARK: - properties
-    var coordinator: CreateTimerSetViewCoordinator!
+    var coordinator: LocalTimeSetViewCoordinator!
     
     // MARK: - lifecycle
     override func loadView() {
-        self.view = CreateTimerSetView()
+        self.view = LocalTimeSetView()
     }
     
     override func viewDidLoad() {
@@ -28,13 +28,13 @@ class CreateTimerSetViewController: BaseViewController, View {
     }
     
     // MARK: - reactor bind
-    func bind(reactor: CreateTimerSetViewReactor) {
+    func bind(reactor: LocalTimeSetViewReactor) {
         // MARK: action
         
         // MARK: state
     }
     
     deinit {
-        Logger.verbose()
+        Logger.verbose("")
     }
 }
