+++ timer/Resources/Assets.xcassets/local.imageset/local.pdf
