+++ timer/Resources/Assets.xcassets/share.imageset/share.pdf
