+++ timer/Support/Info.plist
