--- timer/View/CreateTimerSetView.swift
@@ -1,39 +1,0 @@
-//
-//  CreateTimerSetView.swift
-//  timer
-//
-//  Created by JSilver on 19/05/2019.
-//  Copyright © 2019 Jeong Jin Eun. All rights reserved.
-//
-
-import UIKit
-
-class CreateTimerSetView: UIView {
-    // MARK: - view properties
-    let view: Header = {
-        let view = Header()
-        return view
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        backgroundColor = Constants.Color.white
-        
-        addAutolayoutSubview(view)
-        
-        view.snp.makeConstraints { make in
-            if #available(iOS 11.0, *) {
-                make.top.equalTo(safeAreaLayoutGuide)
-            } else {
-                make.top.equalToSuperview()
-            }
-            
-            make.leading.equalToSuperview()
-            make.trailing.equalToSuperview()
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
